pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "product-cli-grpc"
version = "0.1.0"
authors = [ { name="SIXMAR729", email="webdevsixmar@gmail.com" } ]
description = "A CLI tool to manage a product catalog via a gRPC API."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [ "grpcio >= 1.62.0" ]

[project.scripts]
product-cli = "product_cli.cli:main"